+++ packages/surveycto/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @openfn/language-surveycto
 
+## 2.1.0
+
+### Minor Changes
+
+- b67f81be: Add a `cursor()` function which adds support for the surveyCTO
+  string format dates.
+
+  The date option in `fetchSubmissions` can now accept a surveyCTO date, an
+  epoch or unix timestamp, or basically anything you can pass into new Date().
+
+  Internally, any date you pass into `cursor` will be converted into a
+  human-readable surveyCTO format in UTC time.
+
+### Patch Changes
+
+- Updated dependencies [88f99a8f]
+  - @openfn/language-common@1.13.3
+
 ## 2.0.0
 
 ### Major Changes